# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "選課系統DEMO API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "/api",
            "description": "開發環境API"
        },
        {
            "url": "https://demo-course-system.elf-run.com/api",
            "description": "demo環境API"
        }
    ],
    "paths": {
...
        "/teacher/courses": {
            "post": {
                "tags": [
                    "Teacher"
                ],
                "summary": "新增課程",
                "description": "新增課程",
                "operationId": "createCourse",
                "requestBody": {
                    "description": "課程資料",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "課程名稱",
                                        "type": "string",
                                        "example": "範例課程"
                                    },
                                    "description": {
                                        "description": "課程描述",
                                        "type": "string",
                                        "example": "範例課程敘述"
                                    },
                                    "content": {
                                        "description": "課程大綱",
                                        "type": "string",
                                        "example": "範例課程內容"
                                    },
                                    "college_id": {
                                        "description": "學院ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "department_id": {
                                        "description": "系所ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "level_code": {
                                        "description": "課程等級代碼",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "classroom_id": {
                                        "description": "教室ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "credit": {
                                        "description": "學分數",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "is_required": {
                                        "description": "是否必修",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "semester_code": {
                                        "description": "學年度代碼",
                                        "type": "string",
                                        "example": "115-1"
                                    },
                                    "max_students": {
                                        "description": "人數上限",
                                        "type": "integer",
                                        "example": 30
                                    },
                                    "remarks": {
                                        "description": "備註",
                                        "type": "string",
                                        "example": "無"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "課程創建成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "description": "狀態代碼",
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "description": "結果描述",
                                            "type": "string",
                                            "example": "課程創建成功"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "課程ID",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "description": "課程名稱",
                                                    "type": "string",
                                                    "example": "範例課程"
                                                },
                                                "description": {
                                                    "description": "課程描述",
                                                    "type": "string",
                                                    "example": "範例課程敘述"
                                                },
                                                "is_required": {
                                                    "description": "是否必修",
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "teach_id": {
                                                    "description": "授課教師ID",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "max_students": {
                                                    "description": "人數上限",
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "now_students": {
                                                    "description": "目前人數",
                                                    "type": "integer",
                                                    "example": 0
                                                },
                                                "created_at": {
                                                    "description": "建立時間",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2021-07-07 12:00:00"
                                                },
                                                "updated_at": {
                                                    "description": "更新時間",
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2021-07-07 12:00:00"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "responses": {
            "401": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "description": "狀態",
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "description": "訊息",
                                    "type": "string",
                                    "example": "Validation Error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "422": {
                "description": "Validation Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "description": "狀態",
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "description": "訊息",
                                    "type": "string",
                                    "example": "Validation Error"
                                },
                                "errors": {
                                    "description": "錯誤訊息",
                                    "properties": {
                                        "message": {
                                            "description": "錯誤訊息",
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The message field is required."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}